1_0_build_install_64.ps1 - rake bin files

+++ 1_0_build_install_64.ps1
@@ -364,13 +364,17 @@ Remove-Item Env:\SOURCE_DATE_EPOCH
 
 # Fix rake bin files
 
-$content = [IO.File]::ReadAllText("$d_install/bin/racc", $UTF8)
-$content = $content -replace 'racc', 'rake'
-[IO.File]::WriteAllText("$d_install/bin/rake", $content, $UTF8)
+if (Test-Path -Path "$d_install/bin/racc" -PathType Leaf ) {
+  $content = [IO.File]::ReadAllText("$d_install/bin/racc", $UTF8)
+  $content = $content -replace 'racc', 'rake'
+  [IO.File]::WriteAllText("$d_install/bin/rake", $content, $UTF8)
+}
 
-$content = [IO.File]::ReadAllText("$d_install/bin/racc.cmd", $UTF8)
-$content = $content -replace 'racc', 'rake'
-[IO.File]::WriteAllText("$d_install/bin/rake.cmd", $content, $UTF8)
+if (Test-Path -Path "$d_install/bin/racc.cmd" -PathType Leaf ) {
+  $content = [IO.File]::ReadAllText("$d_install/bin/racc.cmd", $UTF8)
+  $content = $content -replace 'racc', 'rake'
+  [IO.File]::WriteAllText("$d_install/bin/rake.cmd", $content, $UTF8)
+}
 
 if (Test-Path -Path "$d_install/bin/rake.bat" -PathType Leaf ) {
   Remove-Item -Path "$d_install/bin/rake.bat"
